# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wolf IPTV Admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stats {
  totalAccounts: number;
  activeAccounts: number;
  xtreamCount: number;
  m3uCount: number;
  cacheActiveCount: number;
  cacheExpiredCount: number;
}

export type AccountType = typeof AccountType[keyof typeof AccountType];


export const AccountType = {
  xtream: 'xtream',
  m3u: 'm3u',
} as const;

export interface Account {
  id: number;
  type: AccountType;
  name: string;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  user?: string | null;
  /** @nullable */
  pass?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  avatar?: string | null;
  active?: boolean;
  /** @nullable */
  cacheUpdatedAt?: string | null;
  createdAt: string;
}

export type AccountInputType = typeof AccountInputType[keyof typeof AccountInputType];


export const AccountInputType = {
  xtream: 'xtream',
  m3u: 'm3u',
} as const;

export interface AccountInput {
  type: AccountInputType;
  name: string;
  host?: string;
  user?: string;
  pass?: string;
  url?: string;
  avatar?: string;
  active?: boolean;
}

export interface AccountUpdate {
  name?: string;
  host?: string;
  user?: string;
  pass?: string;
  url?: string;
  avatar?: string;
  active?: boolean;
}

export interface CacheStatus {
  accountId: number;
  hasLive: boolean;
  hasMovies: boolean;
  hasSeries: boolean;
  /** @nullable */
  lastUpdated: string | null;
  isExpired: boolean;
  /** @nullable */
  hoursRemaining?: number | null;
}