# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wolf IPTV Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountType } from './accountType';

export interface Account {
  id: number;
  type: AccountType;
  name: string;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  user?: string | null;
  /** @nullable */
  pass?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  avatar?: string | null;
  active?: boolean;
  /** @nullable */
  cacheUpdatedAt?: string | null;
  createdAt: string;
}