# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wolf IPTV Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInputType } from './accountInputType';

export interface AccountInput {
  type: AccountInputType;
  name: string;
  host?: string;
  user?: string;
  pass?: string;
  url?: string;
  avatar?: string;
  active?: boolean;
}