# lib/api-zod/src/generated/types/accountUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wolf IPTV Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountUpdate {
  name?: string;
  host?: string;
  user?: string;
  pass?: string;
  url?: string;
  avatar?: string;
  active?: boolean;
}