# lib/api-zod/src/generated/types/cacheStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wolf IPTV Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface CacheStatus {
  accountId: number;
  hasLive: boolean;
  hasMovies: boolean;
  hasSeries: boolean;
  /** @nullable */
  lastUpdated: string | null;
  isExpired: boolean;
  /** @nullable */
  hoursRemaining?: number | null;
}