# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wolf IPTV Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalAccounts: number;
  activeAccounts: number;
  xtreamCount: number;
  m3uCount: number;
  cacheActiveCount: number;
  cacheExpiredCount: number;
}